vagrant/catalog/requirements.txt
requests
oauth2client
Flask==0.10.1
SQLAlchemy==0.8.4
WTForms==2.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Flask-Mail==0.9.1
Flask-Login==0.3.2
Flask-Rauth
Werkzeug==0.11.4
bcrypt==2.0.0
Flask-Bcrypt==0.7.1
python-slugify